Move SecondaryActionUiModel to package 'nativeapm'
